refactor: config table in several files
