+++ apps/front/src/components/coaches/table/config/email.tsx
@@ -1,0 +1,20 @@
+import { Button } from "@/components/ui/button"
+import { Coach } from "@/types/coach"
+import { CaretSortIcon } from "@radix-ui/react-icons"
+import { ColumnDef } from "@tanstack/react-table"
+
+export const emailColumn: ColumnDef<Coach> = {
+  accessorKey: "email",
+  header: ({ column }) => {
+    return (
+      <Button
+        variant="ghost"
+        onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
+      >
+        Email
+        <CaretSortIcon className="ml-2 h-4 w-4" />
+      </Button>
+    )
+  },
+  cell: ({ row }) => <div className="lowercase">{row.original.email}</div>,
+}
