+++ apps/front/src/components/coaches/table/config/select.tsx
@@ -1,0 +1,26 @@
+import { Checkbox } from "@/components/ui/checkbox";
+import { Coach } from "@/types/coach";
+import { ColumnDef } from "@tanstack/react-table";
+
+export const selectColumn: ColumnDef<Coach> = {
+  id: "select",
+  header: ({ table }) => (
+    <Checkbox
+      checked={
+        table.getIsAllPageRowsSelected() ||
+        (table.getIsSomePageRowsSelected() && "indeterminate")
+      }
+      onCheckedChange={(value) => table.toggleAllPageRowsSelected(!!value)}
+      aria-label="Select all"
+    />
+  ),
+  cell: ({ row }) => (
+    <Checkbox
+      checked={row.getIsSelected()}
+      onCheckedChange={(value) => row.toggleSelected(!!value)}
+      aria-label="Select row"
+    />
+  ),
+  enableSorting: false,
+  enableHiding: false,
+}
