+++ apps/front/src/types/coach.ts
@@ -1,0 +1,9 @@
+export type Coach = {
+  id: number;
+  lastName: string;
+  firstName: string;
+  email: string;
+  phone: string;
+  picture?: string;
+  numberOfCustomers?: number;
+};
